move UiTestInitializer into UiTest annotation
